Implement "Adding to Favorites" by coin Swiping menu
