+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/category/MarketCategoryFragment.kt
@@ -127,7 +127,12 @@ fun CategoryScreen(
                                 }
 
                                 viewItems?.let {
-                                    CoinList(it, scrollToTopAfterUpdate, onCoinClick)
+                                    CoinList(
+                                        it,
+                                        scrollToTopAfterUpdate,
+                                        { uid -> viewModel.onToggleFavorite(uid) },
+                                        onCoinClick
+                                    )
                                     if (scrollToTopAfterUpdate) {
                                         scrollToTopAfterUpdate = false
                                     }
