+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/favorites/MarketFavoritesToggleService.kt
@@ -1,0 +1,26 @@
+package io.horizontalsystems.bankwallet.modules.market.favorites
+
+import io.horizontalsystems.bankwallet.core.managers.MarketFavoritesManager
+
+class MarketFavoritesToggleService(
+    private val favoritesManager: MarketFavoritesManager
+) {
+
+    val allFavorites: HashSet<String>
+        get() = favoritesManager.getAll()
+            .map { it.coinUid }
+            .toHashSet()
+
+    fun toggleCoinFavorite(uid: String) {
+        val inFavorites = favoritesManager.isCoinInFavorites(uid)
+        if (inFavorites) {
+            favoritesManager.remove(uid)
+        } else {
+            favoritesManager.add(uid)
+        }
+    }
+
+    fun remove(uid: String) {
+        favoritesManager.remove(uid)
+    }
+}
