+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/topcoins/MarketTopCoinsViewModel.kt
@@ -9,6 +9,7 @@ import io.horizontalsystems.bankwallet.core.subscribeIO
 import io.horizontalsystems.bankwallet.entities.DataState
 import io.horizontalsystems.bankwallet.entities.ViewState
 import io.horizontalsystems.bankwallet.modules.market.*
+import io.horizontalsystems.bankwallet.modules.market.favorites.MarketFavoritesToggleService
 import io.horizontalsystems.bankwallet.modules.market.topcoins.MarketTopCoinsModule.Menu
 import io.horizontalsystems.bankwallet.ui.compose.Select
 import io.reactivex.disposables.CompositeDisposable
@@ -17,6 +18,7 @@ import kotlinx.coroutines.launch
 
 class MarketTopCoinsViewModel(
     private val service: MarketTopCoinsService,
+    private val favoritesToggleService: MarketFavoritesToggleService,
     private var marketField: MarketField
 ) : ViewModel() {
 
@@ -81,9 +83,10 @@ class MarketTopCoinsViewModel(
     }
 
     private fun syncMarketViewItems() {
+        val favorites = favoritesToggleService.allFavorites
         viewItemsLiveData.postValue(
             marketItems.map {
-                MarketViewItem.create(it, marketField)
+                MarketViewItem.create(it, marketField, favorites.contains(it.fullCoin.coin.uid))
             }
         )
     }
@@ -138,5 +141,10 @@ class MarketTopCoinsViewModel(
         selectorDialogStateLiveData.postValue(
             SelectorDialogState.Opened(Select(service.sortingField, service.sortingFields))
         )
+    }
+
+    fun onToggleFavorite(uid: String) {
+        favoritesToggleService.toggleCoinFavorite(uid)
+        syncMarketViewItems()
     }
 }
