+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/walletconnect/list/ui/ActionRow.kt
@@ -1,41 +1,19 @@
 package io.horizontalsystems.bankwallet.modules.walletconnect.list.ui
 
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
-import io.horizontalsystems.bankwallet.R
 
 @Composable
 fun ActionsRow(
-    actionIconWidth: Dp,
-    onDelete: () -> Unit,
+    content: @Composable () -> Unit,
 ) {
     Row(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp),
+        modifier = Modifier.fillMaxSize(),
         horizontalArrangement = Arrangement.End
     ) {
-        IconButton(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(actionIconWidth),
-            onClick = {
-                onDelete()
-            },
-            content = {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_circle_minus_24),
-                    tint = Color.Gray,
-                    contentDescription = "delete",
-                )
-            }
-        )
+        content()
     }
 }
