+++ components/icons/src/main/res/drawable/ic_star_off_24.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12.9133,2.5927C12.7525,2.2322 12.3947,2 12,2C11.6053,2 11.2475,2.2322 11.0867,2.5927L9.7099,5.6801C9.4849,6.1845 9.7115,6.7758 10.2159,7.0007C10.7203,7.2256 11.3115,6.9991 11.5365,6.4947L12,5.4553L13.8404,9.5821C13.986,9.9086 14.2948,10.1325 14.6504,10.1694L19.1506,10.6368L18.1728,11.5123C17.7613,11.8807 17.7264,12.5129 18.0948,12.9244C18.4632,13.3358 19.0954,13.3707 19.5069,13.0023L22.1671,10.6204C22.4619,10.3564 22.5733,9.9435 22.4513,9.5671C22.3293,9.1907 21.9969,8.9216 21.6033,8.8807L15.4316,8.2398L12.9133,2.5927Z"
+      android:fillColor="#808085"/>
+  <path
+      android:pathData="M18.0138,17.1354L18.8462,20.7772C18.9348,21.1648 18.7854,21.568 18.4656,21.8043C18.1459,22.0405 17.7165,22.0649 17.3721,21.8664L12,18.7709L6.628,21.8664C6.2856,22.0637 5.8593,22.0409 5.5399,21.8083C5.2206,21.5757 5.0681,21.1768 5.1509,20.7905L6.4451,14.7501L1.8329,10.6204C1.5382,10.3564 1.4267,9.9435 1.5487,9.5671C1.6707,9.1907 2.0031,8.9216 2.3967,8.8807L6.4468,8.4601L2.9,5.8C2.4582,5.4686 2.3686,4.8418 2.7,4.4C3.0314,3.9582 3.6582,3.8686 4.1,4.2L22.1,17.7C22.5418,18.0314 22.6314,18.6582 22.3,19.1C21.9686,19.5418 21.3418,19.6314 20.9,19.3L18.0138,17.1354ZM8.7615,10.1962C8.6813,10.2269 8.5958,10.2477 8.5064,10.257L4.8494,10.6368L8.2116,13.6472C8.4789,13.8866 8.5975,14.2509 8.5223,14.6017L7.5776,19.011L11.5007,16.7503C11.8098,16.5723 12.1902,16.5723 12.4993,16.7503L16.3862,18.9901L15.5378,15.2784L8.7615,10.1962Z"
+      android:fillColor="#808085"
+      android:fillType="evenOdd"/>
+</vector>
